+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [Unreleased] - 2019-11-08
+### Added
+- Control function wrappers
+
 ## [0.0.2-alpha] - 2019-11-04
 ### Version
 - Bump for release